requirements.txt
absolufy-imports==0.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.4.0
Authlib==1.6.5
beautifulsoup4==4.14.2
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
cryptography==46.0.3
cyclopts==3.24.0
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.2
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.118.3
fastmcp==2.12.5
filelock==3.20.0
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2025.9.0
google-adk==1.16.0
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.121.0
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.38.0
google-cloud-bigtable==2.33.0
google-cloud-core==2.4.3
google-cloud-discoveryengine==0.13.12
google-cloud-logging==3.12.1
google-cloud-monitoring==2.28.0
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.25.0
google-cloud-spanner==3.58.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.45.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.4
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
idna==3.11
importlib_metadata==8.7.0
inflection==0.5.1
ipykernel==7.0.1
ipython==8.37.0
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
langchain-core==1.0.0
langchain-mcp-adapters==0.1.11
langsmith==0.4.37
lazy-object-proxy==1.12.0
litellm==1.66.3
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mcp==1.16.0
mdurl==0.1.2
more-itertools==10.8.0
multidict==6.7.0
multitasking==0.0.12
nest-asyncio==1.6.0
numpy==2.2.6
openai==2.5.0
openapi==2.0.0
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-logging==1.10.0a0
opentelemetry-exporter-gcp-monitoring==1.10.0a0
opentelemetry-exporter-gcp-trace==1.10.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.10.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
packaging==25.0
pandas==2.3.3
parse==1.20.2
parso==0.8.5
pathable==0.4.4
peewee==3.18.2
platformdirs==4.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.5
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.27.1
rsa==4.9.1
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sqlalchemy-spanner==1.17.0
sqlparse==0.5.3
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
tenacity==8.5.0
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.3.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.1
yarl==1.22.0
yfinance==0.2.56
zipp==3.23.0
zstandard==0.25.0
google-adk[a2a]